refactor(skill-progression): extract Layer 5 view functions into _views.py (#72)
Move get_skill_profile, get_skill_timeline, get_skill_audit, and
get_avg_skill_level_checkpoints verbatim from skill_progression_service.py
into app/services/skill_progression/_views.py (Layer 5).

- _views.py imports calculate_tournament_skill_contribution explicitly
  from ._ema_engine (Layer 4 symbol interaction)
- All repeated logic blocks (TEAM/INDIVIDUAL total_players branch,
  player_baseline_avg pattern, dual calculate_skill_value_from_placement
  calls in get_skill_audit) preserved verbatim — no deduplication
- __init__.py updated: docstring → "Layers 1+2+3+4+5", Layer 5 import
  block, 4 new symbols added to __all__
- shim now a pure 52-line re-export file; orphaned header imports
  (math, User, TournamentSkillMapping, etc.) left as pre-existing debt
- Test patch paths updated for all 18 affected methods:
  _BASE_VIEWS + 4 _PATCH_*_VIEWS constants added; TestGetSkillTimeline
  (5 methods), TestGetSkillAudit (6 methods), TestSkillProgressionBranchBuffer
  E+F sub-sections (5 methods) now target _views.__dict__
- 69/69 unit tests pass, 2490/2490 integration tests pass (0 regressions)
