+++ app/services/skill_progression/__init__.py
@@ -1,12 +1,12 @@
 """
-skill_progression sub-package — Layers 1 + 2 + 3 + 4.
+skill_progression sub-package — Layers 1 + 2 + 3 + 4 + 5.
 
 Layer 1 (_formulas):    pure math, constants
 Layer 2 (_config):      skill key enumeration, baseline lookup, tournament mapping
 Layer 3 (_db_helpers):  DB-backed opponent factor + match performance modifier
 Layer 4 (_ema_engine):  sequential EMA history-replay loops (ctsc + cstsd)
+Layer 5 (_views):       per-user skill profile, timeline, audit, and checkpoint views
 
-Future layers (views) will be added in subsequent PRs.
 The canonical import surface remains app.services.skill_progression_service (thin shim).
 """
 from ._formulas import (
@@ -30,6 +30,12 @@ from ._ema_engine import (
     calculate_tournament_skill_contribution,
     compute_single_tournament_skill_delta,
 )
+from ._views import (
+    get_skill_profile,
+    get_skill_timeline,
+    get_skill_audit,
+    get_avg_skill_level_checkpoints,
+)
 
 __all__ = [
     "MIN_SKILL_VALUE",
@@ -45,4 +51,8 @@ __all__ = [
     "_compute_match_performance_modifier",
     "calculate_tournament_skill_contribution",
     "compute_single_tournament_skill_delta",
+    "get_skill_profile",
+    "get_skill_timeline",
+    "get_skill_audit",
+    "get_avg_skill_level_checkpoints",
 ]
